gio: Fix various enum -Wsign-conversion warnings
The `g_enum_*()` APIs operate purely on `int` (because they need to be
polymorphic), but match based on the binary value rather than the
integer value, so signed/unsigned conversions don’t affect the results.

Fix various `-Wsign-conversion` warnings with them by making the
implicit sign conversion explicit.


Helps: #3405

+++ gio/gcredentials.c
@@ -406,7 +406,7 @@ credentials_native_type_check (GCredentialsType  requested_type,
 #endif
 
   enum_class = g_type_class_ref (g_credentials_type_get_type ());
-  requested = g_enum_get_value (enum_class, requested_type);
+  requested = g_enum_get_value (enum_class, (int) requested_type);
 
 #if G_CREDENTIALS_SUPPORTED
   supported = g_enum_get_value (enum_class, G_CREDENTIALS_NATIVE_TYPE);
